+++ src/commands/rewrite.ts
@@ -17,15 +17,31 @@ export const rewriteArgs = z.object({
   /**
    * Remove original snapshots after creating new ones
    */
-  forget: z.coerce.boolean(),
+  forget: z.coerce.boolean().default(false),
   /**
    * Exclude pattern(s) but ignore case in patterns
    */
-  iexcludePattern: z.coerce.string().array().default([]),
+  iexclude: z.coerce.string().array().default([]),
   /**
    * Read exclude patterns from given file(s) but ignore case in patterns
    */
   iexcludeFile: z.coerce.string().array().default([]),
+  /**
+   * Include pattern(s)
+   */
+  include: z.string().array().default([]),
+  /**
+   * Read include patterns from given file(s)
+   */
+  includeFile: z.string().array().default([]),
+  /**
+   * Include pattern(s) but ignore case in patterns
+   */
+  iinclude: z.coerce.string().array().default([]),
+  /**
+   * Read include patterns from given file(s) but ignore case in patterns
+   */
+  iincludeFile: z.coerce.string().array().default([]),
   /**
    * Replace hostname
    */
@@ -37,7 +53,7 @@ export const rewriteArgs = z.object({
   /**
    * Create snapshot summary record if it does not exist
    */
-  snapshotSummary: z.coerce.boolean(),
+  snapshotSummary: z.coerce.boolean().default(false),
 });
 
 class RewriteArgumentBuilder extends RepositoryArgumentBuilder<string, string> {